requirements.txt
asn1crypto==0.24.0
cached-property==1.5.1
certifi==2018.10.15
chardet==3.0.4
crypto==1.4.1
idna==2.7
Naked==0.1.31
python-pkcs11==0.5.0
PyYAML==3.13
requests==2.20.0
shellescape==3.4.1
urllib3==1.24
